Matrix images looks static
The displayed pattern looks now static. It uses almost a frequency of
1MHz, the internal clock frequency, to push into the shift register.
_delay_us(0.001) was too slow. But it works fine.

With a frequency of 1MHz the period is 1µs - more than enough to meet
the timing requirements of the shift register.

f = 1,000,000 -> 1/f = 0.000,001 = 1µs

Maybe I should implement it with a timer ISR, just for the purpose of
power consumption.